Organize modules into app and tests packages
