+++ app/silence_detector.py
